+++ src/common/message/ShapeData.java
@@ -8,9 +8,9 @@ import java.io.Serializable;
 public class ShapeData implements Serializable {
     public final Point focus;
     public final Drawable[] shapes;
-    public final int clientShipHealth;
+    public final double clientShipHealth;
 
-    public ShapeData(Point focus, Drawable[] shapes, int clientShipHealth) {
+    public ShapeData(Point focus, Drawable[] shapes, double clientShipHealth) {
         this.focus = focus;
         this.shapes = shapes;
         this.clientShipHealth = clientShipHealth;
